Fixing ClangTidy issues reported by downstream integration
ClangTidy reported 16 issues. All are around typecasting and
straightforward

+++ vp9/encoder/vp9_temporal_filter.c
@@ -424,8 +424,8 @@ void vp9_highbd_apply_temporal_filter_c(
     }
   }
 
-  for (row = 0; row < (int)uv_block_height; row++) {
-    for (col = 0; col < (int)uv_block_width; col++) {
+  for (row = 0; row < uv_block_height; row++) {
+    for (col = 0; col < uv_block_width; col++) {
       const int u_diff =
           u_src[row * uv_src_stride + col] - u_pre[row * uv_pre_stride + col];
       const int v_diff =
@@ -480,8 +480,8 @@ void vp9_highbd_apply_temporal_filter_c(
   }
 
   // Apply the filter to chroma
-  for (uv_row = 0; uv_row < (int)uv_block_height; uv_row++) {
-    for (uv_col = 0; uv_col < (int)uv_block_width; uv_col++) {
+  for (uv_row = 0; uv_row < uv_block_height; uv_row++) {
+    for (uv_col = 0; uv_col < uv_block_width; uv_col++) {
       const int y_row = uv_row << ss_y;
       const int y_col = uv_col << ss_x;
       const int filter_weight = get_filter_weight(
